Split out classes from fuddly.framework.knowledge.information to fuddly.info.generic
