+++ pyproject.toml
@@ -45,3 +45,6 @@ tuto = "fuddly.data_models.tutorial.tuto"
 standard = "fuddly.projects.generic.standard"
 usb = "fuddly.projects.specific.usb"
 tuto = "fuddly.projects.tuto"
+
+[project.entry-points."fuddly.info"]
+generic = "fuddly.info.generic"
